requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
python-http-client==3.3.7
selenium==4.9.1
sendgrid==6.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
wsproto==1.2.0
